Drop instanceHint from VintfObject::ListInstances.
ListInstances used to be a function that represents either
a manifest or an IServiceManager. In the case that it represents
a passthrough service manager, the service manager cannot
list instance names directly, so it needs the matrix to tell
the passthrough service manager what instance names are expected.

With I78d4699fcecb0d54
("Refactor vts_treble_vintf_test"),
DeprecateTest.NoDeprecatedHalsOnManager was deleted. We no
longer needed that test because we already enforce:
- No deprecate HALs must be present on manifests
- All HALs on the service manager must be declared in manifests
  (and all HALs in the manifest must be retrievable via
   service managers).

Because we only check deprecation on manifests,
instanceHint is no longer needed to get all instance names.
Hence, we can safely drop this argument for simpler code.

Test: TH
Bug: 280314899
Change-Id: Icb7b279872d82d4fbe40ab885d9e4b229fbe6c53