Merge VintfObject::checkCompatibility.
Now that we no longer needs an abstract
ListInstances function to represent a HAL manifest
or a service manager, simply replace ListInstances
with the device manifest.

Remove the override of checkCompatibility with
ListInstances because the device manifest is used
directly.

This change prepares for AIDL deprecation check, where
the AIDL instances are needed from the device manifest
as well.

Test: TH
Bug: 280314899
Change-Id: I9e6632fb522b743b564b43bd3c07530bf63ad7fc