Fix AIDL versions in checkMissingHalsInMatrices.
Previously when AIDL HALs are checked in checkMissingHalsInMatrices,
versions doesn't exist in AIDL HALs. Now that AIDL HALs have versions,
properly check the versions in matrices.

libaidlmetadata lists only frozen versions in the .versions
field. When has_development is true, also add the next version
to the list so the next version is required to be in compatibility
matrices (or excluded).

Test: TH
Bug: 255383566
Change-Id: I3fa7a0ede59d89a87bf3c55ca260823f8e8d7b97