Remove new unfrozen HAL manifest entries on 'next'
If a service declares V1 of a new unfrozen AIDL interface, it is removed
from the manifest with RELEASE_AIDL_USE_UNFROZEN is false (in the 'next'
release configuration).
This causes these services to fail to register themselves with service
manager.
If the service is in a process by itself, it should not be installed on
the device when RELEASE_AIDL_USE_UNFROZEN is false.
If the service is in a process that must be installed on the device, the
service should check with service manager if it is declared
(IServiceManager::isDeclared()) before attempting to register itself.

Test: atest vts_treble_vintf_vendor_test
Bug: 292151564
Change-Id: I233a81a0f3a204bfbf4da1919f4af941dc06b2ea
Merged-In: I233a81a0f3a204bfbf4da1919f4af941dc06b2ea