Improve deprecation message for vendor extensions.
If vendor.foo@1.0 inherits from android.hardware.foo@1.0,
and the latter is deprecated, the error message was

  vendor.foo@1.0 is deprecated ... it should not be served.
  android.hardware.foo@1.0 is deprecated ... it should not be served.

However, if one just look at the first line, they might
be confused why vendor.foo@1.0 is deprecated. Add the reason text.

New message:
  android.hardware.foo@1.0/... is deprecated [...] it should not be served.
    because it matches android.hardware.foo@1.0 from /system/etc/vintf/compatibility_matrix.<level>.xml
vendor.foo@1.0/... is deprecated [...] it should not be served.
    because it inherits from vendor.foo@1.0/... that matches android.hardware.foo@1.0/... from /system/etc/vintf/compatibility_matrix.<level>.xml


Bug: 295238794
Test: TH
Change-Id: I35ec7df944a87bf3a7dc116865e6a51c99ec38d1