Use unsigned arg for HealthHalVersion ctor
HealthHalVersion is a typedef for std::variant<Version, size_t>, and
Version is a struct representing a pair of size_t values.

After upgrading libc++, implicit conversion of an int to a
HealthHalVersion no longer works, presumably because libc++ has
implemented the P0608R3 fix. Replace the "1" int argument with a "1u"
unsigned int argument.

Bug: 175635923
Test: treehugger
Test: m MODULES-IN-system-libvintf-test
Change-Id: If1ff84308eee804fdf7b6d1da3f2d86f3169c811