<matrix><hal> optional attr has default value true.
If a compatibility-matrix.hal does not
specify attribute value for the attribute
optional, the default value changes from
false to true.

This changes the meaning for new VINTF
files on old libvintf, hence metaversion
is bumped.

All FCMs in hardware/interfaces has already
switched to optional=true.

All device FCMs (device-specific FCMs in the
/system, /system_ext and /product) uses optional
as a mechanism to check if a HAL is installed,
but it is disassociated from the modules that would
actually have the dependency (there is no
fcm_vintf_fragments).

All DCMs (matrices in /vendor, /odm) are not
tested and inaccurate, so it is okay to switch
the default.

After this change, all FCMs in hardware/interfaces
can drop optional=true and use the implicit default
value.

Also bump metaversion to 8.0, indicating older libvintf
to reject XML files at version 8.0 which uses true as
the default value for optional.

Bug: 247575800
Test: TH
Change-Id: I5b1a842fcbdca2a90aa5df3755b0a67c50317c87