Add next year's release Level
This is needed for to support devices that are launching next year.
The numbers and letters used for the release are just placeholders for
right now and may be changed in the future.

The flag protects the user of this change. This change only allows the
new behavior to be supported.
Flag: build.RELEASE_AIDL_USE_UNFROZEN
Test: atest -p
Bug: 346861728

Change-Id: I6a1840b93a573d4bfa18863b12a0a4317d007d05