hals_for_release.py: ignore some FCMs
We only need to consider the standard FCMs in the format
compatibilty_matrix.<level>.xml.
Ignore the compatibility_matrix.<level>.android17.xml file.

Now that we are deprecating the old FCMs earlier in the release cycle,
we have added a new version of the old to-be-deprecated FCM in order to
deprecate them for trunk* configs while keeping them around for the next
configs for releasing devices that are still treated as the previous
desert version.

Test: python hals_for_release.py
Bug: 415071628

Change-Id: I10738f72fea6ac0c0d2efb45d4dcd44c73068376